backend/requirements.txt
fastapi
uvicorn[standard]
sqlmodel
asyncpg
pydantic
pydantic-settings
httpx
redis[hiredis]
python-multipart

# dev
pytest
pytest-asyncio

# email
google-auth
google-auth-oauthlib
google-api-python-client
